auralis/requirements.txt
pysyncobj
numpy
pytest
flwr
fastapi
uvicorn
websockets
torch
torchvision
pyyaml
requests
opacus
shap
opencv-python
matplotlib
scipy
httpx
